infrastructure 0.16.5: change bitnami registry to legacy
in the infrastructure chart bitnami charts and with that bitnami docker
images are used for rabbitmq, postgresql, and redis
These dependencies now use the 'bitnamilegacy' docker repository

bitnami has changed the license and release of docker images.
https://news.broadcom.com/app-dev/broadcom-introduces-bitnami-secure-images-for-production-ready-containerized-applications

The images used in the chart dependency will no longer be developed by
bitnami.
Existing images are moved to the 'bitnamilegacy' docker repository.